refactor: extract PullRequestCommentService into dedicated files
